ml/requirements.txt
Flask
tensorflow
Pillow
numpy
flask_cors